Container Class Added
Added a container class that can now be nested (visually) inside another
container
How the container code needs to be written so the child is inside it,
not the parent
Moved the event handler to the Mode class
Moved all drawing functions to the render function in the Window class
Removed the border drawing function.  A new one will likely be needed in
the future, but I'll cross that bridge when I need to
Changed load function to create all necessary screens at once. May be
changed, but we'll see